--- TicTacToe/README.txt
@@ -1,14 +1,0 @@
-TicTacToe
-
--- An old game.
-
-How to Run It
-=============
-
-$ javac ./*.java
-$ java Main
-
-How to Play It
-==============
-
-Please search it.
